+++ NEWS.md
@@ -11,7 +11,8 @@
 * Faster RNG generator for Laplace distribution
 * Changes to using lower level RNG functions
   (`unif_rand`, `norm_rand`) when possible
-* More tests for checking multivariate distributions
+* More tests
+* Bernoulli pmf now does not warn if x is not 0 or 1
 
 ### 1.7.2